+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `report-exit-code` action. Exits with exit code 1 if duplicates found, 0 otherwise.
 - `quiet` output format. Suppresses all output, useful for scripting in combination with `report-exit-code` action.
+- `--max-size` filter to skip files larger than a given size.
 
 ## [0.1.1] - 2025-12-30